fakejob detector/requirements.txt
flask
joblib
scikit-learn
pandas
beautifulsoup4
textblob
googlesearch-python
gunicorn

